Implemented MovieClip and started implementing TilingSprite (works on canvas but not yet webgl). Fixed ordering issues and transparency issues in the webgl renderer. Converted more of the examples to pixi.dart

+++ lib/src/core/matrix.dart
@@ -22,7 +22,7 @@ abstract class Matrix
 
 	void operator []=(int index, num value)
 	{
-		this._source[index] = value;
+		this._source[index] = value.toDouble();
 	}
